third_party/WebKit/LayoutTests/imported/wpt/dom/nodes/MutationObserver-inner-outer-expected.txt

This is a testharness.js-based test.
FAIL innerHTML mutation assert_equals: mutation records must match expected 2 but got 1
PASS innerHTML with 2 children mutation 
PASS outerHTML mutation 
Harness: the test ran to completion.
